+++ CHANGES.rst
@@ -5,6 +5,7 @@ Changelog
 ------------------
 
 - fixed access of `output_sampledata` flag of `wai.spectralio.adams.Writer`
+- added setter for sample data in `Spectrum` class
 
 
 0.0.1 (2025-04-14)